separating ui element from app.js & reuse planning
